修改 mesh 索引
修改 mesh 索引

+++ Creator0.10/Source/Runtime/CrMeshUtility.cpp
@@ -116,7 +116,7 @@ CrMesh *CrMeshUtility::CreateMeshCube()
 		1, 0,
 	};
 
-	GLushort triangles[] =
+	GLuint triangles[] =
 	{
 		0, 2, 3, 0, 3, 1, 8, 4, 5, 8, 5, 9, 10, 6, 7, 10, 7, 11, 12, 13, 14, 12, 14, 15, 16, 17, 18, 16, 18, 19, 20, 21, 22, 20, 22, 23,
 	};
@@ -224,7 +224,7 @@ CrMesh *CrMeshUtility::CreateMeshQuad()
 		1.0f, 1.0f,
 	};
 
-	GLushort triangles[] =
+	GLuint triangles[] =
 	{
 		0, 1, 3,
 		0, 3, 2,
